Splitted installers into different files
