+++ generamba/lib/generamba/template/helpers/rambaspec_validator.rb
