+++ generamba/lib/generamba/template/installer/abstract_installer.rb
@@ -1,0 +1,7 @@
+module Generamba
+  class AbstractInstaller
+    def install_template(template_declaration)
+      raise 'Abstract Method - you should implement it in the concrete subclass'
+    end
+  end
+end
