+++ generamba/lib/generamba/template/installer/remote_installer.rb
@@ -1,0 +1,9 @@
+require 'generamba/template/installer/abstract_installer.rb'
+
+module Generamba
+  class RemoteInstaller < AbstractInstaller
+    def install_template(template_declaration)
+      puts('Still not implemented')
+    end
+  end
+end
