+++ generamba/lib/generamba/template/processor/template_declaration.rb
